Save game (#4)
Add core methods for saving and loading games

+++ FrigidRogue.MonoGame.Core/Services/SaveGameStore.cs
@@ -1,7 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Reflection;
+
+using AutoMapper;
 
 using FrigidRogue.MonoGame.Core.Components;
 using FrigidRogue.MonoGame.Core.Interfaces.Components;
@@ -10,6 +13,7 @@ using FrigidRogue.MonoGame.Core.Interfaces.Services;
 using MonoGame.Framework.Utilities.Deflate;
 
 using Newtonsoft.Json;
+using NGenerics.Extensions;
 
 namespace FrigidRogue.MonoGame.Core.Services
 {
@@ -18,12 +22,14 @@ namespace FrigidRogue.MonoGame.Core.Services
         private readonly JsonSerializerSettings _jsonSerializerSettings;
 
         private Dictionary<Type, string> _jsonObjectStore = new Dictionary<Type, string>();
+        public IMapper Mapper { get; set; }
 
         public SaveGameStore()
         {
             _jsonSerializerSettings = new JsonSerializerSettings
             {
-                ObjectCreationHandling = ObjectCreationHandling.Replace
+                ObjectCreationHandling = ObjectCreationHandling.Replace,
+                TypeNameHandling = TypeNameHandling.All
             };
         }
 
@@ -34,6 +40,18 @@ namespace FrigidRogue.MonoGame.Core.Services
             return new Memento<T>(JsonConvert.DeserializeObject<T>(jsonString, _jsonSerializerSettings));
         }
 
+        private T GetFromStore<T>(Dictionary<Type, string> jsonStore)
+        {
+            var key = jsonStore.Keys.FirstOrDefault(k => typeof(ILoadGameDetail).IsAssignableFrom(k));
+
+            if (key == null)
+                throw new Exception($"An object was not found in the store which is or can be assigned as a type {typeof(T)}");
+
+            var jsonString = jsonStore[key];
+
+            return (T)JsonConvert.DeserializeObject(jsonString, key, _jsonSerializerSettings);
+        }
+
         public void SaveToStore<T>(IMemento<T> memento)
         {
             var jsonString = JsonConvert.SerializeObject(memento.State, _jsonSerializerSettings);
@@ -41,7 +59,45 @@ namespace FrigidRogue.MonoGame.Core.Services
             _jsonObjectStore.Add(typeof(T), jsonString);
         }
 
-        public void SaveStoreToFile(string saveGameName)
+        public T GetFromStore<T, TSaveData>()
+        {
+            var saveData = GetFromStore<TSaveData>().State;
+
+            return Mapper.Map<TSaveData, T>(saveData);
+        }
+
+        public T GetFromStore<T, TSaveData>(T existingObject)
+        {
+            var saveData = GetFromStore<TSaveData>().State;
+
+            return Mapper.Map(saveData, existingObject);
+        }
+
+        public void SaveToStore<T, TSaveData>(T item)
+        {
+            var saveData = Mapper.Map<T, TSaveData>(item);
+
+            SaveToStore(new Memento<TSaveData>(saveData));
+        }
+
+        public void Clear()
+        {
+            _jsonObjectStore = new Dictionary<Type, string>();
+        }
+
+        public SaveGameResult CanSaveStoreToFile(string saveGameName)
+        {
+            var saveGameFolder = GetSaveGamePath();
+
+            var saveGameFile = Path.Combine(saveGameFolder, $"{saveGameName}.sav");
+
+            if (File.Exists(saveGameFile))
+                return SaveGameResult.Overwrite;
+            else
+                return SaveGameResult.Success;
+        }
+
+        public SaveGameResult SaveStoreToFile(string saveGameName, bool overwrite)
         {
             var saveGameFolder = GetSaveGamePath();
 
@@ -51,20 +107,57 @@ namespace FrigidRogue.MonoGame.Core.Services
 
             var saveGameBytes = GZipStream.CompressString(saveGameString);
 
-            File.WriteAllBytes(saveGameFile, saveGameBytes);
+            if (!overwrite && File.Exists(saveGameFile))
+                return SaveGameResult.Overwrite;
+
+            try
+            {
+                File.WriteAllBytes(saveGameFile, saveGameBytes);
+            }
+            catch (Exception e)
+            {
+                var messageTemplate = $"Error when saving game to filesystem. Filename: {saveGameFile}";
+
+                Logger.Error(e, messageTemplate);
+                return new SaveGameResult { ErrorMessage = $"{messageTemplate} - {e.Message}" };
+            }
+
+            return SaveGameResult.Success;
         }
 
-        public void LoadStoreFromFile(string saveGameName)
+        public LoadGameResult LoadStoreFromFile(string saveGameName)
         {
             var saveGameFolder = GetSaveGamePath();
 
+            try
+            {
+                var deserialisedStore = DeserialiseStoreFromFile(saveGameName, saveGameFolder);
+
+                _jsonObjectStore = deserialisedStore;
+            }
+            catch (Exception e)
+            {
+                var message = $"Could not load game {saveGameName}";
+                Logger.Error(e, message);
+
+                return new LoadGameResult { ErrorMessage = $"{message}: {e.Message}" };
+            }
+
+            return LoadGameResult.Success;
+        }
+
+        private Dictionary<Type, string> DeserialiseStoreFromFile(string saveGameName, string saveGameFolder)
+        {
             var saveGameFile = Path.Combine(saveGameFolder, $"{saveGameName}.sav");
 
             var saveGameBytes = File.ReadAllBytes(saveGameFile);
 
             var saveGameString = GZipStream.UncompressString(saveGameBytes);
 
-            _jsonObjectStore = JsonConvert.DeserializeObject<Dictionary<Type, string>>(saveGameString, _jsonSerializerSettings);
+            var deserialisedStore =
+                JsonConvert.DeserializeObject<Dictionary<Type, string>>(saveGameString, _jsonSerializerSettings);
+
+            return deserialisedStore;
         }
 
         public string GetSaveGamePath()
@@ -81,6 +174,55 @@ namespace FrigidRogue.MonoGame.Core.Services
                 Directory.CreateDirectory(saveGameFolder);
 
             return saveGameFolder;
+        }
+
+        public IList<LoadGameDetails> GetLoadGameList()
+        {
+            var saveGamePath = GetSaveGamePath();
+
+            var directoryInfo = new DirectoryInfo(saveGamePath);
+
+            var fileInfos = directoryInfo.GetFiles("*.sav");
+
+            var loadGameList = new List<LoadGameDetails>(fileInfos.Length);
+
+            foreach (var fileInfo in fileInfos)
+            {
+                Dictionary<Type, string> deserialisedFile;
+
+                try
+                {
+                    deserialisedFile = DeserialiseStoreFromFile(Path.GetFileNameWithoutExtension(fileInfo.Name), saveGamePath);
+                }
+                catch (Exception e)
+                {
+                    Logger.Warning(e, $"Exception occurred when deserialising file {fileInfo.Name}");
+                    continue;
+                }
+
+                ILoadGameDetail loadGameDetails;
+
+                try
+                {
+                    loadGameDetails = GetFromStore<ILoadGameDetail>(deserialisedFile);
+                }
+                catch (Exception e)
+                {
+                    Logger.Warning(e, $"Exception occurred when getting ILoadGameDetail from save file {fileInfo.Name}");
+                    continue;
+                }
+
+                loadGameList.Add(
+                    new LoadGameDetails
+                    {
+                        DateTime = fileInfo.LastWriteTime,
+                        Filename = Path.GetFileNameWithoutExtension(fileInfo.Name),
+                        LoadGameDetail = loadGameDetails.LoadGameDetail
+                    }
+                );
+            }
+
+            return loadGameList;
         }
     }
 }
